working on the Service layer
